checkpoint: this is a trap
