+++ signer.js
@@ -1,1 +1,11 @@
 // implements abstract signer class
+async function getAbstractSigner(signer, verbose = true) {
+	// TODO: create abstract signer class that is compatible with ethers v5, v6, viem, web3js
+	return signer
+}
+
+
+/* This is a compatibility stub, to enable module referencing in commonJS */
+module.exports = {
+    getAbstractSigner
+}
